feature/CU-jx9k5x_BACK-END-Add-quiz-creation-service

+++ src/services/db/db.js
@@ -42,22 +42,11 @@ class quizDB {
     * @param collection String, collection to drop, Required
     */
     async dropCollection(collection) {
-        let collectionsCursor = await this.db.listCollections();
-        if( (await collectionsCursor.toArray()).indexOf(collection) >= 0 ){
-            await this.db.dropCollection(collection);
-        }  
-    }
+        let collectionsCursor = await this.db.listCollections({}, { nameOnly: true });
+        let collections = await collectionsCursor.toArray();
+        let collectionNames = collections.map(c =>c.name);
 
-    /**
-    * Create collection
-    *
-    * @param collection String, collection to create, Required
-    */
-    async createCollection(collection) {
-        let collectionsCursor = await this.db.listCollections();
-        if( (await collectionsCursor.toArray()).indexOf(collection) < 0 ){
-            await this.db.createCollection(collection);
-        } 
+        if (collectionNames.indexOf(collection) >= 0) await this.db.dropCollection(collection);
     }
 
     /**
